Remove draft proposals for now
