+++ templates/spend.tsx
@@ -43,7 +43,7 @@ export const SpendComponent = ({
 
   return (
     <div className="flex justify-between items-center bg-base-300 py-2 px-3 rounded-lg my-2">
-      <div className="flex items-center gap-x-4 gap-y-2 flex-wrap">
+      <div className="flex items-center gap-4 flex-wrap">
         <div className="flex items-center flex-wrap gap-x-2 gap-y-2 w-24">
           <h2 className="text-4xl">💵</h2>
           <h2>Spend</h2>
